Develop

+++ config/Debug/Debug.php
@@ -1,6 +1,10 @@
 <?php
 
-function dump($variable, ...$variables)
+/**
+ * @param mixed $variable
+ * @param mixed $variables
+ */
+function dump($variable, ...$variables): void
 {
     $data = [$variable];
 
@@ -17,14 +21,18 @@ function dump($variable, ...$variables)
     }
 }
 
-function dd($variable, ...$variables)
+/**
+ * @param mixed $variable
+ * @param mixed $variables
+ */
+function dd($variable, ...$variables): void
 {
     dump($variable, ...$variables);
 
     die();
 }
 
-function deleteTwigCacheFolder()
+function deleteTwigCacheFolder(): void
 {
     $dir = dirname(__DIR__, 2).'\var\\cache\\twig';
     if (is_dir($dir)) {
@@ -32,22 +40,28 @@ function deleteTwigCacheFolder()
     }
 }
 
-function deleteDir($dirPath)
+function deleteDir(string $dirPath): void
 {
     if (! is_dir($dirPath)) {
         throw new InvalidArgumentException("$dirPath must be a directory");
     }
+    
     if (substr($dirPath, strlen($dirPath) - 1, 1) != '/') {
         $dirPath .= '/';
     }
+
     $files = glob($dirPath . '*', GLOB_MARK);
-    foreach ($files as $file) {
-        if (is_dir($file)) {
-            deleteDir($file);
-        } else {
-            unlink($file);
+
+    if (is_array($files)) {
+        foreach ($files as $file) {
+            if (is_dir($file)) {
+                deleteDir($file);
+            } else {
+                unlink($file);
+            }
         }
     }
+
     rmdir($dirPath);
 }
